+++ packages/admin-sdk/src/message-types.ts
@@ -1,6 +1,6 @@
 import type { notificationDispatch } from './notification/index';
 import type { toastDispatch } from './toast';
-import type { windowRedirect, windowReload, windowRouterPush } from './window/index';
+import type { windowRedirect, windowReload, windowRouterPush, windowGetId } from './window/index';
 import type {
   contextLanguage,
   contextEnvironment,
@@ -51,6 +51,7 @@ export interface ShopwareMessageTypes {
   windowRedirect: windowRedirect,
   windowRouterPush: windowRouterPush,
   windowReload: windowReload,
+  windowGetId: windowGetId,
   contextLanguage: contextLanguage,
   contextEnvironment: contextEnvironment,
   contextLocale: contextLocale,