+++ examples/admin-sdk-plugin/src/Resources/app/administration/CHANGELOG.md
@@ -1,5 +1,26 @@
 # administration
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [13ce32a]
+- Updated dependencies [6372237]
+- Updated dependencies [55832a9]
+  - @shopware-ag/meteor-admin-sdk@5.7.7
+  - @shopware-ag/meteor-component-library@4.5.1
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [ee0a4ad]
+- Updated dependencies [6de3f71]
+- Updated dependencies [a57d22e]
+- Updated dependencies [d988d78]
+  - @shopware-ag/meteor-component-library@4.5.0
+  - @shopware-ag/meteor-admin-sdk@5.7.6
+
 ## 1.0.8
 
 ### Patch Changes